Working version with equal sign
The nasty little = is couted as a COMP_WORDBREAK.

But you can't unset that because that breaks all other scripts
(since it's an env variable).

So the solution is to use _get_comp_words_by_ref -n "="
where -n is exclude character.

So $cur is now --color=<blah>
instead of $cur="<blah>" and $prev="=" (since = is a word boundary,
curr will be the "new" word instead of the entire string.