Works except git adds two spaces on its completions
Merge the local completion with _git_diff (when
no -l is present).

Problems:
Git adds two spaces to its completions.
Find out why, and then use sed if it can't be
removed by programatically.

It should stop completing locally on the
first unknown local parameter and only
return git completions from that point on.

Find out how to check if a parameter is not
on a list and if so return only git completions,
otherwise merge.

-l kills any further local completion, which
is not correct. It should stop and only return
_git_log when the first unknown is hit.

Same as above, but then -l is present on the
line instead.